factory_vec: Add unit tests
The tests use GTK, which must be initialized and run on a single
designated thread. By default, tests run on multiple threads and will
fail. As a workaround, the total number of threads can be set to one
with `cargo test -- --test-threads=1`.

Implemented with macros from the `relm4-macros` crate, which is listed
as an optional dependency in `Cargo.toml`. In lack of an alternative,
the crate is added as a dev dependency.