Prevent leaking CommandSenderInner struct
Earlier, this struct was being leaked due to the implementation of
Deref. However, since the struct was never actually exported,
documentation was not being generated for it's method.

This commit just makes the struct private and instead creates wrapper
functions over it. This is typically whats done in Rust std.

Signed-off-by: Ayush Singh <ayushsingh1325@gmail.com>