Add method `broadcast` to `FactoryVecDeque`
In order to send a message to all elements in `FactoryVecDeque`,
it was previously necessary to manually cycle through all the
indices (from 0 to `len()`) and send message to each element, one by
one, using an existing `send` method. This new method makes this
process cleaner and safer by avoiding the need to use panic-prone
indices.