Fix nightly Rust version for "docs" CI job to prevent errors regarding "doc_auto_cfg" removal.
"doc_auto_cfg" became part of "doc_cfg" in nightly and targeted to be
stable in 1.92. As such docs generation for all the crates that rely on
it fails with nightly version where "doc_auto_cfg" was removed. As this
was unstable flag all the dependencies has to migrate from "doc_auto_cfg" to
"doc_cfg". Once Relm4 upgrade all the dependencies to "fixed" versions
we can lift nightly restriction.

Documentation:
- https://doc.rust-lang.org/rustdoc/unstable-features.html#doc_auto_cfg-automatically-generate-doccfg
- https://github.com/rust-lang/rust/issues/43781

Changes:
- Set nightly Rust version to "2025-09-27" for "docs" CI job in
  .github/workflows/rust.yml as this is the latest nightly version that
  works.